Create PseRespondentViewController
