+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/controllers/PseRespondToTribunalController.java
@@ -76,40 +76,6 @@ public class PseRespondToTribunalController {
     }
 
     /**
-     *  Populates the dynamic list for select an order or request to respond to.
-     *
-     * @param ccdRequest holds the request and case data
-     * @param userToken  used for authorization
-     * @return Callback response entity with case data attached.
-     */
-    @PostMapping(value = "/aboutToStartView", consumes = APPLICATION_JSON_VALUE)
-    @Operation(summary = "Populates the dynamic list for Select a judgment, order or notification")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Accessed successfully",
-            content = {
-                @Content(mediaType = "application/json",
-                    schema = @Schema(implementation = CCDCallbackResponse.class))
-            }),
-        @ApiResponse(responseCode = "400", description = "Bad Request"),
-        @ApiResponse(responseCode = "500", description = "Internal Server Error")
-    })
-    public ResponseEntity<CCDCallbackResponse> aboutToStartView(
-        @RequestBody CCDRequest ccdRequest,
-        @RequestHeader(value = "Authorization") String userToken) {
-
-        if (!verifyTokenService.verifyTokenSignature(userToken)) {
-            log.error(INVALID_TOKEN, userToken);
-            return ResponseEntity.status(FORBIDDEN.value()).build();
-        }
-
-        CaseData caseData = ccdRequest.getCaseDetails().getCaseData();
-        caseData.setPseRespondentSelectJudgmentOrderNotification(
-            pseRespondToTribunalService.populateSelectDropdownView(caseData));
-
-        return getCallbackRespEntityNoErrors(caseData);
-    }
-
-    /**
      * Middle Event for initial Request/Order details.
      * @param ccdRequest        CaseData which is a generic data type for most of the
      *                          methods which holds case data
