+++ src/test/apiTest/resources/application.properties
@@ -1,5 +1,5 @@
-idam.url=https://idam-api.aat.platform.hmcts.net
-docmosis.test.url=http://et-cos.service.core-compute-aat.internal
+idam.url=http://localhost:5000
+docmosis.test.url=http://localhost:8081
 
 spring.datasource.url=jdbc:postgresql://localhost:5050/et_cos
